+++ arch/Sparc/SparcDisassembler.c
@@ -212,7 +212,7 @@ static DecodeStatus readInstruction32(const uint8_t *code, size_t len, uint32_t 
 	*Insn = (code[3] <<  0) |
 		(code[2] <<  8) |
 		(code[1] << 16) |
-		(code[0] << 24);
+		((uint32_t) code[0] << 24);
 
 	return MCDisassembler_Success;
 }
